+++ spec/rubocop/cop/lint/void_spec.rb
@@ -15,18 +15,14 @@ RSpec.describe RuboCop::Cop::Lint::Void do
         a %{op} b
       RUBY
     end
-  end
 
-  described_class::BINARY_OPERATORS.each do |op|
     it "accepts void op #{op} if on last line" do
       expect_no_offenses(<<~RUBY)
         something
         a #{op} b
       RUBY
     end
-  end
 
-  described_class::BINARY_OPERATORS.each do |op|
     it "accepts void op #{op} by itself without a begin block" do
       expect_no_offenses("a #{op} b")
     end
@@ -67,9 +63,7 @@ RSpec.describe RuboCop::Cop::Lint::Void do
         #{op}b
       RUBY
     end
-  end
 
-  unary_operators.each do |op|
     it "accepts void unary op #{op} by itself without a begin block" do
       expect_no_offenses("#{op}b")
     end